EpochCollection now contains a notion for the "latest"epoch
LatestEpoch -1 now triggers an error, because it produces inconsistent
results
Improved debug information and timing about precommitting and comitting
events
The waiting source now waits for last epoch +1 to prevent the commit on
the sink from not being called in some edge cases

+++ src/main/scala/org/codefeedr/model/zookeeper/EpochCollection.scala
@@ -1,3 +1,3 @@
 package org.codefeedr.model.zookeeper
 
-case class EpochCollection (latestEpoch: Int)
+case class EpochCollection(latestEpoch: Long)
